pyproject.toml
[tool.poetry]
name = "langchain-demo"
version = "0.1.0"
description = "LangChain & prompt engineering demo setup"
package-mode = false
authors = ["Prasad Ovhal"]


[tool.poetry.dependencies]
python = "3.11.*"

# Core ML
torch = "*"
transformers = "*"
datasets = "*"
evaluate = "*"
accelerate = "*"
timm = "*"

# LLM / HF ecosystem (let transformers manage internals)
safetensors = "*"

# LangChain ecosystem
langchain = "*"
langchain-core = "*"
langchain-community = "*"
langchain-openai = "*"
langchain-google-genai = "*"
google-generativeai = "*"

# Validation
pydantic = "^2"

[tool.poetry.group.dev.dependencies]
ipykernel = "*"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"